+++ requake/formulas/__init__.py
@@ -9,3 +9,5 @@ Formulas for the Requake package.
     GNU General Public License v3.0 or later
     (https://www.gnu.org/licenses/gpl-3.0-standalone.html)
 """
+from .moment import mag_to_moment  # noqa
+from .slip import mag_to_slip_in_cm # noqa